add routes by API docs
