+++ routes/choice.js
@@ -1,12 +1,6 @@
 const express = require(`express`);
 const router = express.Router();
 // API 명세를 보고 여기부터 코드 작성해주세요.
-router.route(`/`) // http://localhost:4000/register
-    .post((request, response) => {
-        return;
-    })
-    .get((request, response) => {
-        return;
-    })
+
 // 여기까지
 module.exports = router;
