+++ routes/vote.js
@@ -1,0 +1,22 @@
+const express = require(`express`);
+const router = express.Router();
+
+router.route(`/`)
+    .post((req, res) => {
+        
+    })    
+    .get((req, res) => {
+        
+    })
+    .put((req, res) => {
+        
+    })
+    .patch((req, res) => {
+        
+    })
+    .delete((req, res) => {
+        
+    });
+
+    
+module.exports = router;
